--- src/components/List/ListDetail.jsx
@@ -1,6 +1,0 @@
-import React from 'react';
-import CommonListDetail from './CommonListDetail';
-
-export default function ListDetail() {
-  return <CommonListDetail />;
-}
